Let bundler >= 2.2 handle windows platform
Fixes #1157

Because it has support for multiple platforms in Gemfile.lock. So deleting Gemfile.lock when one of the platforms is windows is not good.
Bundler's error message tells the user how to handle it.
```
Your bundle only supports platforms ["mingw"] but your local platform is x86_64-linux. Add the current platform to the lockfile with
`bundle lock --add-platform x86_64-linux` and try again.
```

+++ lib/language_pack/ruby.rb
@@ -693,7 +693,7 @@ BUNDLE
         WARNING
       end
 
-      if bundler.windows_gemfile_lock?
+      if !bundler.supports_multiple_platforms? && bundler.windows_gemfile_lock?
         log("bundle", "has_windows_gemfile_lock")
 
         File.unlink("Gemfile.lock")
